refactor(shadow): organize actions/functions
