+++ src/components/shadow/umbra.tsx
@@ -2,7 +2,9 @@ import { createRenderEffect, createSignal, For, onMount, Show } from "solid-js";
 import { createStore } from "solid-js/store";
 import { isTest } from "../../actions/test-actions";
 import ShadowEl from "./shadow-el";
-import { scaleAndCenterRect, scaleAndCenterVec, ShadowOriginOptions, ShadowRect, UmbraState, ZERO_RECT } from "./types";
+import { ShadowOriginOptions, ShadowRect, UmbraState, ZERO_RECT } from "./types";
+import { scaleAndCenterVec } from "~/actions/vector-actions";
+import { scaleAndCenterRect } from "./actions";
 
 export interface UmbraProps {
   /**
